+++ CHANGELOG.md
@@ -2,12 +2,22 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [38.11.0] 2026-01-16
+## [40.1.0] 2026-01-22
 ### Fixed
 - A Questionnaire with multiple initial values on a repeating question
   should render multiple questions, each with a single default answer.
 - The same form, when exported back to Questionnaire, should contain
   only one question with an array of initial values.
+
+## [40.0.0] 2026-01-22
+### Added
+- Support for observationExtractEntry extension in Observation extract.
+- Observation-based extraction will now always return a Bundle.
+
+## [39.0.0] 2026-01-20
+### Changed
+- Throw an error if duplicate FHIRPath variable names are found
+  on the same Questionnaire item or root.
 
 ## [38.10.1] 2026-01-13
 ### Changed